pyproject.toml
[project]
name = "vipani-sutra"
version = "0.1.0"
description = "Hyperlocal listings API (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.10.0",
    "cloudinary>=1.41.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "firebase-admin>=6.5.0",
]

[tool.uv]
package = false
